Refactor: rework code to be more OO
The old C pattern has been reworked.
A Parser class has been introduced.
His goal is to get the command and call the correct method.
It owns a shared_ptr to the second class Debugger.
The Debugger class owns the differents info from the current programm.
It implements the different command handler method

Signed-off-by: magnar_c <clement.magnard@epita.fr>